# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate an image from a text prompt
 */
export const GenerateGeminiImageBody = zod.object({
  prompt: zod.string(),
});

export const GenerateGeminiImageResponse = zod.object({
  b64_json: zod.string(),
  mimeType: zod.string(),
});

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});